uboot-mvebu: disable the internal i2c slave on i2c0
This patch removes the internal i2c slave on Armada 38x i2c0 bus.
Problems with the i2c bus starts to appear when working with i2c0 bus on
speeds higher than 100khz (bus locks etc...).

Note: This slave is only on i2c0 and is not connected to i2c1.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>